Challenge: Sprite pachinko
It’s time to put your new skills to the test! In this challenge assignment, you will create a pachinko-style interaction using Sprites. You will make a circle (or whatever you like) fall through a grid of “pins” (other Sprites) to see where it lands.
To complete the requirements for this Mission, submit your 2D project to the gallery in this tutorial. (You may share it with others or keep it private.)